requirements.txt
flask==2.3.3
flask-login==0.6.2
supabase==1.0.4
python-dotenv==1.0.0
bcrypt==4.1.2
